requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
eventlet==0.33.3
extra-streamlit-components==0.1.71
firebase-admin==6.7.0
Flask==2.2.3
Flask-Compress==1.13
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
frozenlist==1.5.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.80.0
google-auth==2.38.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.2.0
narwhals==1.32.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.68.2
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyairtable==3.0.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.0
pydantic-settings==2.0.2
pydantic_core==2.27.0
pydeck==0.9.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.1
Pyrebase4==4.8.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.11.2
python-jwt==4.1.0
python-socketio==5.12.1
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
ratelimit==2.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
setuptools==78.0.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.43.2
streamlit-authenticator==0.2.1
streamlit-cookies-manager==0.2.0
streamlit-javascript==0.1.5
streamlit-login-auth-ui==0.2.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
tenacity==8.2.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trycourier==6.2.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uuid==1.30
webencodings==0.5.1
Werkzeug==2.2.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.18.3
